+++ tests/test_config.py
@@ -35,6 +35,7 @@ class ConfigTests(unittest.TestCase):
             "user_agent",
             "ignore_attribute",
             "allow_dynamic_classes",
+            "send_exception_details",
         ):
             self.assertEqual(getattr(config1, member), getattr(config2, member))
 
@@ -63,6 +64,7 @@ class ConfigTests(unittest.TestCase):
         config1.serialize_method = "_new_method"
         config1.ignore_attribute = "_new_method"
         config1.allow_dynamic_classes = True
+        config1.send_exception_details = True
         self.compare_config(config1, config1.copy())
 
         # Handlers